refactor(planning): retire epic-planner.js and inline its exports into epic-plan-spec.js (resolves #1446)
