+++ .agents/scripts/epic-plan-spec.js
@@ -29,13 +29,13 @@
 import { readFile } from 'node:fs/promises';
 import path from 'node:path';
 import { parseArgs } from 'node:util';
-import { buildAuthoringContext, planEpic } from './epic-planner.js';
 import {
   commitSnapshotsToEpicBranch,
   forkMainToEpic,
 } from './lib/baseline-snapshot.js';
 import { runAsCli } from './lib/cli-utils.js';
 import {
+  getLimits,
   PROJECT_ROOT,
   resolveConfig,
   validateOrchestrationConfig,
@@ -45,15 +45,204 @@ import * as gitUtils from './lib/git-utils.js';
 import { Logger } from './lib/Logger.js';
 import { AGENT_LABELS, TYPE_LABELS } from './lib/label-constants.js';
 import { PlanRunnerContext } from './lib/orchestration/context.js';
+import { buildDocsContext } from './lib/orchestration/doc-reader.js';
 import {
   PLAN_PHASES,
   PlanCheckpointer,
 } from './lib/orchestration/plan-runner/plan-checkpointer.js';
 import { sweepStaleStoryWorktrees } from './lib/orchestration/plan-runner/worktree-sweep.js';
+import { applyBudget } from './lib/orchestration/planning-context-budget.js';
+import { PlanningStateManager } from './lib/orchestration/planning-state-manager.js';
 import { cleanupPhaseTempFiles } from './lib/plan-phase-cleanup.js';
 import { createProvider } from './lib/provider-factory.js';
 import { forceDrainPendingCleanup } from './lib/worktree/lifecycle/force-drain.js';
 import { readManifest } from './lib/worktree/lifecycle/pending-cleanup.js';
+
+// ─── PRD / Tech Spec system prompts ──────────────────────────────────────────
+//
+// These are the canonical authoring prompts that ride along on the
+// `--emit-context` envelope as a backstop. The `epic-plan-spec-author` Skill
+// (`.agents/skills/core/epic-plan-spec-author/SKILL.md`) embeds the
+// authoritative copies of these strings — keep the two surfaces in sync when
+// either is edited.
+
+export const PRD_SYSTEM_PROMPT = `You are an expert Technical Product Manager.
+Your job is to convert a high-level Epic description into a structured Product Requirements Document (PRD).
+
+The PRD should outline:
+1. Context & Goals
+2. User Stories
+3. Acceptance Criteria
+4. Out of Scope
+
+CRITICAL REQUIREMENTS:
+- Respond ONLY with valid Markdown.
+- Do not use top-level <h1> (# ) tags. Start with ## Overview.
+- Format requirements clearly with bullet points and bold text where appropriate.`;
+
+export const TECH_SPEC_SYSTEM_PROMPT = `You are an expert Engineering Architect.
+Your job is to convert a PRD into a Technical Specification for implementation.
+
+The Tech Spec should outline:
+1. Architecture & Design
+2. Data Models (if any)
+3. API Changes (if any)
+4. Core Components
+5. Security & Privacy Considerations
+
+CRITICAL REQUIREMENTS:
+- Respond ONLY with valid Markdown.
+- Do not use top-level <h1> (# ) tags. Start with ## Technical Overview.
+- Format architectural decisions clearly with bullet points.`;
+
+/**
+ * Build the authoring context the host LLM (or the
+ * `epic-plan-spec-author` Skill) needs to write the PRD and Tech Spec.
+ *
+ * Returns a plain JSON-serialisable object; never hits the network beyond the
+ * provider call needed to load the Epic.
+ *
+ * `docsContext` is bounded by the planning-context budget (Epic #817 Story 9):
+ * over-budget payloads downgrade to a summary representation with headings +
+ * bounded excerpts. Pass `{ fullContext: true }` (CLI: `--full-context`) to
+ * restore the unbounded full-body envelope. The Epic body itself is always
+ * subject to the same budget so a sprawling Epic narrative cannot bypass the
+ * cap by riding on top of `docsContext`.
+ */
+export async function buildAuthoringContext(
+  epicId,
+  provider,
+  settings = {},
+  opts = {},
+) {
+  const epic = await provider.getEpic(epicId);
+  if (!epic) {
+    throw new Error(`Epic #${epicId} not found.`);
+  }
+
+  const planningLimits = getLimits({ agentSettings: settings }).planningContext;
+  const { fullContext = false } = opts;
+
+  const docsContext = await buildDocsContext(settings, planningLimits, {
+    fullContext,
+  });
+
+  const epicBody = applyBudget(
+    [{ path: `epic-${epic.id}.md`, content: epic.body ?? '' }],
+    planningLimits,
+    { fullContext },
+  );
+
+  return {
+    epic: {
+      id: epic.id,
+      title: epic.title,
+      body: epicBody.mode === 'full' ? epic.body : null,
+      bodySummary: epicBody.mode === 'summary' ? epicBody.items[0] : null,
+      linkedIssues: epic.linkedIssues ?? { prd: null, techSpec: null },
+    },
+    docsContext,
+    systemPrompts: {
+      prd: PRD_SYSTEM_PROMPT,
+      techSpec: TECH_SPEC_SYSTEM_PROMPT,
+    },
+  };
+}
+
+/**
+ * Persist the host-authored PRD and Tech Spec under the Epic.
+ *
+ * Heals any prior planning artifacts (PRD / Tech Spec issues, "Planning
+ * Artifacts" body section, lifecycle labels) before writing the new issues.
+ * Idempotent against partial state: when the Epic already has a PRD but no
+ * Tech Spec, the existing PRD is reused. Pass `force: true` to force a full
+ * regeneration (close prior PRD/Tech Spec and re-create both).
+ */
+export async function planEpic(
+  epicId,
+  provider,
+  { prdContent, techSpecContent },
+  _settings = {},
+  { force = false } = {},
+) {
+  if (typeof prdContent !== 'string' || prdContent.trim() === '') {
+    throw new Error(
+      '[Epic Planner] prdContent is required and must be non-empty.',
+    );
+  }
+  if (typeof techSpecContent !== 'string' || techSpecContent.trim() === '') {
+    throw new Error(
+      '[Epic Planner] techSpecContent is required and must be non-empty.',
+    );
+  }
+
+  Logger.info(`[Epic Planner] Fetching Epic #${epicId}...`);
+  const epic = await provider.getEpic(epicId);
+
+  if (!epic) {
+    throw new Error(`Epic #${epicId} not found.`);
+  }
+
+  const stateManager = new PlanningStateManager(provider);
+  await stateManager.healAndCleanupArtifacts(epic, force);
+
+  // M-8: Resumable planning — if PRD exists but Tech Spec doesn't, resume from PRD.
+  if (!force && epic.linkedIssues?.prd && epic.linkedIssues?.techSpec) {
+    Logger.warn(
+      `[Epic Planner] Epic #${epicId} already has both PRD and Tech Spec. Aborting to prevent duplicates. Use --force to re-plan.`,
+    );
+    return;
+  }
+  const existingPrdId = force ? null : (epic.linkedIssues?.prd ?? null);
+
+  let prdId;
+  if (existingPrdId) {
+    Logger.info(
+      `[Epic Planner] Reusing existing PRD #${existingPrdId}. Skipping PRD creation.`,
+    );
+    prdId = existingPrdId;
+  } else {
+    Logger.info(`[Epic Planner] Creating PRD issue for "${epic.title}"...`);
+    const prdTicket = await provider.createTicket(epicId, {
+      title: `[PRD] ${epic.title}`,
+      body: prdContent,
+      labels: ['context::prd'],
+      dependencies: [],
+    });
+    Logger.info(
+      `[Epic Planner] Created PRD Issue #${prdTicket.id} (${prdTicket.url})`,
+    );
+    prdId = prdTicket.id;
+  }
+
+  Logger.info(
+    `[Epic Planner] Creating Tech Spec issue linking to PRD #${prdId}...`,
+  );
+  const techSpecTicket = await provider.createTicket(epicId, {
+    title: `[Tech Spec] ${epic.title}`,
+    body: techSpecContent,
+    labels: ['context::tech-spec'],
+    dependencies: [prdId],
+  });
+  Logger.info(
+    `[Epic Planner] Created Tech Spec Issue #${techSpecTicket.id} (${techSpecTicket.url})`,
+  );
+
+  Logger.info(
+    `[Epic Planner] Updating Epic #${epicId} with linked documents...`,
+  );
+
+  // Format exactly so getEpic regex /PRD:\s*#\d+/i still catches it efficiently.
+  const appendBody = `\n\n## Planning Artifacts\n- [ ] PRD: #${prdId}\n- [ ] Tech Spec: #${techSpecTicket.id}\n`;
+  const newBody = epic.body + appendBody;
+
+  await provider.updateTicket(epicId, {
+    body: newBody,
+  });
+
+  Logger.info(`[Epic Planner] Epic #${epicId} updated successfully.`);
+  Logger.info(`[Epic Planner] Planning pipeline complete!`);
+}
 
 /**
  * Runs `sweepStaleStoryWorktrees` when a ticketing `provider` is available
