+++ .agents/scripts/lib/orchestration/README.md
@@ -84,7 +84,8 @@ entry point. They never own business logic.
 | Script                               | Purpose                                                                 |
 | ------------------------------------ | ----------------------------------------------------------------------- |
 | `agents-bootstrap-github.js`         | Idempotent setup of GitHub labels and project fields                    |
-| `epic-planner.js`                    | Autonomous PRD and Tech Spec generation                                 |
+| `epic-plan-spec.js`                  | Phase 1 — PRD + Tech Spec context emit + persist (dispatches `epic-plan-spec-author` Skill) |
+| `epic-plan-decompose.js`             | Phase 2 — ticket-array context emit + persist (dispatches `epic-plan-decompose-author` Skill) |
 | `ticket-decomposer.js`               | Recursive 4-tier hierarchy decomposition                                |
 | `dispatcher.js`                      | CLI wrapper — DAG scheduler; outputs dispatch manifest                  |
 | `context-hydrator.js`                | CLI wrapper — assembles self-contained agent prompts                    |
